Rename everything in progress_http to be ingests_api
