+++ storage/ingests_api/src/main/scala/uk/ac/wellcome/platform/storage/ingests/api/Router.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.progress_http
+package uk.ac.wellcome.platform.storage.ingests.api
 
 import java.net.URL
 import java.util.UUID
@@ -7,21 +7,21 @@ import akka.http.scaladsl.marshalling.Marshal
 import akka.http.scaladsl.model._
 import akka.http.scaladsl.model.StatusCodes._
 import akka.http.scaladsl.model.headers.Location
-import akka.stream.scaladsl.Flow
-import akka.util.ByteString
 import akka.http.scaladsl.server.{
   MalformedRequestContentRejection,
   RejectionHandler,
   Route
 }
+import akka.stream.scaladsl.Flow
+import akka.util.ByteString
 import grizzled.slf4j.Logging
 import io.circe.{CursorOp, Printer}
 import uk.ac.wellcome.platform.archive.common.config.models.HTTPServerConfig
+import uk.ac.wellcome.platform.archive.common.models.StorageSpace
 import uk.ac.wellcome.platform.archive.common.models.bagit.{
   BagId,
   ExternalIdentifier
 }
-import uk.ac.wellcome.platform.archive.common.models.StorageSpace
 import uk.ac.wellcome.platform.archive.common.progress.models.Progress
 import uk.ac.wellcome.platform.archive.common.progress.monitor.ProgressTracker
 import uk.ac.wellcome.platform.archive.display.{
@@ -29,7 +29,7 @@ import uk.ac.wellcome.platform.archive.display.{
   RequestDisplayIngest,
   ResponseDisplayIngest
 }
-import uk.ac.wellcome.platform.archive.progress_http.model.ErrorResponse
+import uk.ac.wellcome.platform.storage.ingests.api.model.ErrorResponse
 
 import scala.concurrent.ExecutionContext
