+++ storage/ingests_api/src/test/scala/uk/ac/wellcome/platform/storage/ingests/api/ProgressStarterTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.progress_http
+package uk.ac.wellcome.platform.storage.ingests.api
 
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{FunSpec, Matchers}
